+++ Pool/src/pool/Bola4.java
@@ -10,9 +10,9 @@ public class Bola4 extends Bola{
     @Override
     public void paint(Graphics g, JPanel panel){
         g.setColor(new Color(10, 200, 200)); 
-        g.fillOval(x, y, 30, 30);
+        g.fillOval((int)x, (int)y, 30, 30);
     }
-    public Bola4(int x, int y){
+    public Bola4(double x, double y){
         super(x, y, 10);
     }
 }
