+++ Pool/src/pool/Bola7.java
@@ -10,9 +10,9 @@ public class Bola7 extends Bola{
     @Override
     public void paint(Graphics g, JPanel panel){
         g.setColor(new Color(100, 100, 100)); 
-        g.fillOval(x, y, 30, 30);
+        g.fillOval((int)x, (int)y, 30, 30);
     }
-    public Bola7(int x, int y){
+    public Bola7(double x, double y){
         super(x, y, 10);
     }
 }
